+++ influxdb_client/service/organizations_service.py
@@ -138,118 +138,6 @@ class OrganizationsService(object):
             collection_formats=collection_formats,
             urlopen_kw=urlopen_kw)
 
-    def delete_orgs_id_labels_id(self, org_id, label_id, **kwargs):  # noqa: E501,D401,D403
-        """Delete a label from an organization.
-
-        This method makes a synchronous HTTP request by default. To make an
-        asynchronous HTTP request, please pass async_req=True
-        >>> thread = api.delete_orgs_id_labels_id(org_id, label_id, async_req=True)
-        >>> result = thread.get()
-
-        :param async_req bool
-        :param str org_id: The organization ID. (required)
-        :param str label_id: The label ID. (required)
-        :param str zap_trace_span: OpenTracing span context
-        :return: None
-                 If the method is called asynchronously,
-                 returns the request thread.
-        """  # noqa: E501
-        kwargs['_return_http_data_only'] = True
-        if kwargs.get('async_req'):
-            return self.delete_orgs_id_labels_id_with_http_info(org_id, label_id, **kwargs)  # noqa: E501
-        else:
-            (data) = self.delete_orgs_id_labels_id_with_http_info(org_id, label_id, **kwargs)  # noqa: E501
-            return data
-
-    def delete_orgs_id_labels_id_with_http_info(self, org_id, label_id, **kwargs):  # noqa: E501,D401,D403
-        """Delete a label from an organization.
-
-        This method makes a synchronous HTTP request by default. To make an
-        asynchronous HTTP request, please pass async_req=True
-        >>> thread = api.delete_orgs_id_labels_id_with_http_info(org_id, label_id, async_req=True)
-        >>> result = thread.get()
-
-        :param async_req bool
-        :param str org_id: The organization ID. (required)
-        :param str label_id: The label ID. (required)
-        :param str zap_trace_span: OpenTracing span context
-        :return: None
-                 If the method is called asynchronously,
-                 returns the request thread.
-        """  # noqa: E501
-        local_var_params = locals()
-
-        all_params = ['org_id', 'label_id', 'zap_trace_span']  # noqa: E501
-        all_params.append('async_req')
-        all_params.append('_return_http_data_only')
-        all_params.append('_preload_content')
-        all_params.append('_request_timeout')
-        all_params.append('urlopen_kw')
-
-        for key, val in six.iteritems(local_var_params['kwargs']):
-            if key not in all_params:
-                raise TypeError(
-                    "Got an unexpected keyword argument '%s'"
-                    " to method delete_orgs_id_labels_id" % key
-                )
-            local_var_params[key] = val
-        del local_var_params['kwargs']
-        # verify the required parameter 'org_id' is set
-        if ('org_id' not in local_var_params or
-                local_var_params['org_id'] is None):
-            raise ValueError("Missing the required parameter `org_id` when calling `delete_orgs_id_labels_id`")  # noqa: E501
-        # verify the required parameter 'label_id' is set
-        if ('label_id' not in local_var_params or
-                local_var_params['label_id'] is None):
-            raise ValueError("Missing the required parameter `label_id` when calling `delete_orgs_id_labels_id`")  # noqa: E501
-
-        collection_formats = {}
-
-        path_params = {}
-        if 'org_id' in local_var_params:
-            path_params['orgID'] = local_var_params['org_id']  # noqa: E501
-        if 'label_id' in local_var_params:
-            path_params['labelID'] = local_var_params['label_id']  # noqa: E501
-
-        query_params = []
-
-        header_params = {}
-        if 'zap_trace_span' in local_var_params:
-            header_params['Zap-Trace-Span'] = local_var_params['zap_trace_span']  # noqa: E501
-
-        form_params = []
-        local_var_files = {}
-
-        body_params = None
-        # HTTP header `Accept`
-        header_params['Accept'] = self.api_client.select_header_accept(
-            ['application/json'])  # noqa: E501
-
-        # Authentication setting
-        auth_settings = []  # noqa: E501
-
-        # urlopen optional setting
-        urlopen_kw = None
-        if 'urlopen_kw' in kwargs:
-            urlopen_kw = kwargs['urlopen_kw']
-
-        return self.api_client.call_api(
-            '/api/v2/orgs/{orgID}/labels/{labelID}', 'DELETE',
-            path_params,
-            query_params,
-            header_params,
-            body=body_params,
-            post_params=form_params,
-            files=local_var_files,
-            response_type=None,  # noqa: E501
-            auth_settings=auth_settings,
-            async_req=local_var_params.get('async_req'),
-            _return_http_data_only=local_var_params.get('_return_http_data_only'),  # noqa: E501
-            _preload_content=local_var_params.get('_preload_content', True),
-            _request_timeout=local_var_params.get('_request_timeout'),
-            collection_formats=collection_formats,
-            urlopen_kw=urlopen_kw)
-
     def delete_orgs_id_members_id(self, user_id, org_id, **kwargs):  # noqa: E501,D401,D403
         """Remove a member from an organization.
 
@@ -484,6 +372,9 @@ class OrganizationsService(object):
 
         :param async_req bool
         :param str zap_trace_span: OpenTracing span context
+        :param int offset:
+        :param int limit:
+        :param bool descending:
         :param str org: Filter organizations to a specific organization name.
         :param str org_id: Filter organizations to a specific organization ID.
         :param str user_id: Filter organizations to a specific user ID.
@@ -508,6 +399,9 @@ class OrganizationsService(object):
 
         :param async_req bool
         :param str zap_trace_span: OpenTracing span context
+        :param int offset:
+        :param int limit:
+        :param bool descending:
         :param str org: Filter organizations to a specific organization name.
         :param str org_id: Filter organizations to a specific organization ID.
         :param str user_id: Filter organizations to a specific user ID.
@@ -517,7 +411,7 @@ class OrganizationsService(object):
         """  # noqa: E501
         local_var_params = locals()
 
-        all_params = ['zap_trace_span', 'org', 'org_id', 'user_id']  # noqa: E501
+        all_params = ['zap_trace_span', 'offset', 'limit', 'descending', 'org', 'org_id', 'user_id']  # noqa: E501
         all_params.append('async_req')
         all_params.append('_return_http_data_only')
         all_params.append('_preload_content')
@@ -533,11 +427,23 @@ class OrganizationsService(object):
             local_var_params[key] = val
         del local_var_params['kwargs']
 
+        if 'offset' in local_var_params and local_var_params['offset'] < 0:  # noqa: E501
+            raise ValueError("Invalid value for parameter `offset` when calling `get_orgs`, must be a value greater than or equal to `0`")  # noqa: E501
+        if 'limit' in local_var_params and local_var_params['limit'] > 100:  # noqa: E501
+            raise ValueError("Invalid value for parameter `limit` when calling `get_orgs`, must be a value less than or equal to `100`")  # noqa: E501
+        if 'limit' in local_var_params and local_var_params['limit'] < 1:  # noqa: E501
+            raise ValueError("Invalid value for parameter `limit` when calling `get_orgs`, must be a value greater than or equal to `1`")  # noqa: E501
         collection_formats = {}
 
         path_params = {}
 
         query_params = []
+        if 'offset' in local_var_params:
+            query_params.append(('offset', local_var_params['offset']))  # noqa: E501
+        if 'limit' in local_var_params:
+            query_params.append(('limit', local_var_params['limit']))  # noqa: E501
+        if 'descending' in local_var_params:
+            query_params.append(('descending', local_var_params['descending']))  # noqa: E501
         if 'org' in local_var_params:
             query_params.append(('org', local_var_params['org']))  # noqa: E501
         if 'org_id' in local_var_params:
@@ -678,110 +584,6 @@ class OrganizationsService(object):
             post_params=form_params,
             files=local_var_files,
             response_type='Organization',  # noqa: E501
-            auth_settings=auth_settings,
-            async_req=local_var_params.get('async_req'),
-            _return_http_data_only=local_var_params.get('_return_http_data_only'),  # noqa: E501
-            _preload_content=local_var_params.get('_preload_content', True),
-            _request_timeout=local_var_params.get('_request_timeout'),
-            collection_formats=collection_formats,
-            urlopen_kw=urlopen_kw)
-
-    def get_orgs_id_labels(self, org_id, **kwargs):  # noqa: E501,D401,D403
-        """List all labels for a organization.
-
-        This method makes a synchronous HTTP request by default. To make an
-        asynchronous HTTP request, please pass async_req=True
-        >>> thread = api.get_orgs_id_labels(org_id, async_req=True)
-        >>> result = thread.get()
-
-        :param async_req bool
-        :param str org_id: The organization ID. (required)
-        :param str zap_trace_span: OpenTracing span context
-        :return: LabelsResponse
-                 If the method is called asynchronously,
-                 returns the request thread.
-        """  # noqa: E501
-        kwargs['_return_http_data_only'] = True
-        if kwargs.get('async_req'):
-            return self.get_orgs_id_labels_with_http_info(org_id, **kwargs)  # noqa: E501
-        else:
-            (data) = self.get_orgs_id_labels_with_http_info(org_id, **kwargs)  # noqa: E501
-            return data
-
-    def get_orgs_id_labels_with_http_info(self, org_id, **kwargs):  # noqa: E501,D401,D403
-        """List all labels for a organization.
-
-        This method makes a synchronous HTTP request by default. To make an
-        asynchronous HTTP request, please pass async_req=True
-        >>> thread = api.get_orgs_id_labels_with_http_info(org_id, async_req=True)
-        >>> result = thread.get()
-
-        :param async_req bool
-        :param str org_id: The organization ID. (required)
-        :param str zap_trace_span: OpenTracing span context
-        :return: LabelsResponse
-                 If the method is called asynchronously,
-                 returns the request thread.
-        """  # noqa: E501
-        local_var_params = locals()
-
-        all_params = ['org_id', 'zap_trace_span']  # noqa: E501
-        all_params.append('async_req')
-        all_params.append('_return_http_data_only')
-        all_params.append('_preload_content')
-        all_params.append('_request_timeout')
-        all_params.append('urlopen_kw')
-
-        for key, val in six.iteritems(local_var_params['kwargs']):
-            if key not in all_params:
-                raise TypeError(
-                    "Got an unexpected keyword argument '%s'"
-                    " to method get_orgs_id_labels" % key
-                )
-            local_var_params[key] = val
-        del local_var_params['kwargs']
-        # verify the required parameter 'org_id' is set
-        if ('org_id' not in local_var_params or
-                local_var_params['org_id'] is None):
-            raise ValueError("Missing the required parameter `org_id` when calling `get_orgs_id_labels`")  # noqa: E501
-
-        collection_formats = {}
-
-        path_params = {}
-        if 'org_id' in local_var_params:
-            path_params['orgID'] = local_var_params['org_id']  # noqa: E501
-
-        query_params = []
-
-        header_params = {}
-        if 'zap_trace_span' in local_var_params:
-            header_params['Zap-Trace-Span'] = local_var_params['zap_trace_span']  # noqa: E501
-
-        form_params = []
-        local_var_files = {}
-
-        body_params = None
-        # HTTP header `Accept`
-        header_params['Accept'] = self.api_client.select_header_accept(
-            ['application/json'])  # noqa: E501
-
-        # Authentication setting
-        auth_settings = []  # noqa: E501
-
-        # urlopen optional setting
-        urlopen_kw = None
-        if 'urlopen_kw' in kwargs:
-            urlopen_kw = kwargs['urlopen_kw']
-
-        return self.api_client.call_api(
-            '/api/v2/orgs/{orgID}/labels', 'GET',
-            path_params,
-            query_params,
-            header_params,
-            body=body_params,
-            post_params=form_params,
-            files=local_var_files,
-            response_type='LabelsResponse',  # noqa: E501
             auth_settings=auth_settings,
             async_req=local_var_params.get('async_req'),
             _return_http_data_only=local_var_params.get('_return_http_data_only'),  # noqa: E501
@@ -1434,122 +1236,6 @@ class OrganizationsService(object):
             post_params=form_params,
             files=local_var_files,
             response_type='Organization',  # noqa: E501
-            auth_settings=auth_settings,
-            async_req=local_var_params.get('async_req'),
-            _return_http_data_only=local_var_params.get('_return_http_data_only'),  # noqa: E501
-            _preload_content=local_var_params.get('_preload_content', True),
-            _request_timeout=local_var_params.get('_request_timeout'),
-            collection_formats=collection_formats,
-            urlopen_kw=urlopen_kw)
-
-    def post_orgs_id_labels(self, org_id, label_mapping, **kwargs):  # noqa: E501,D401,D403
-        """Add a label to an organization.
-
-        This method makes a synchronous HTTP request by default. To make an
-        asynchronous HTTP request, please pass async_req=True
-        >>> thread = api.post_orgs_id_labels(org_id, label_mapping, async_req=True)
-        >>> result = thread.get()
-
-        :param async_req bool
-        :param str org_id: The organization ID. (required)
-        :param LabelMapping label_mapping: Label to add (required)
-        :param str zap_trace_span: OpenTracing span context
-        :return: LabelResponse
-                 If the method is called asynchronously,
-                 returns the request thread.
-        """  # noqa: E501
-        kwargs['_return_http_data_only'] = True
-        if kwargs.get('async_req'):
-            return self.post_orgs_id_labels_with_http_info(org_id, label_mapping, **kwargs)  # noqa: E501
-        else:
-            (data) = self.post_orgs_id_labels_with_http_info(org_id, label_mapping, **kwargs)  # noqa: E501
-            return data
-
-    def post_orgs_id_labels_with_http_info(self, org_id, label_mapping, **kwargs):  # noqa: E501,D401,D403
-        """Add a label to an organization.
-
-        This method makes a synchronous HTTP request by default. To make an
-        asynchronous HTTP request, please pass async_req=True
-        >>> thread = api.post_orgs_id_labels_with_http_info(org_id, label_mapping, async_req=True)
-        >>> result = thread.get()
-
-        :param async_req bool
-        :param str org_id: The organization ID. (required)
-        :param LabelMapping label_mapping: Label to add (required)
-        :param str zap_trace_span: OpenTracing span context
-        :return: LabelResponse
-                 If the method is called asynchronously,
-                 returns the request thread.
-        """  # noqa: E501
-        local_var_params = locals()
-
-        all_params = ['org_id', 'label_mapping', 'zap_trace_span']  # noqa: E501
-        all_params.append('async_req')
-        all_params.append('_return_http_data_only')
-        all_params.append('_preload_content')
-        all_params.append('_request_timeout')
-        all_params.append('urlopen_kw')
-
-        for key, val in six.iteritems(local_var_params['kwargs']):
-            if key not in all_params:
-                raise TypeError(
-                    "Got an unexpected keyword argument '%s'"
-                    " to method post_orgs_id_labels" % key
-                )
-            local_var_params[key] = val
-        del local_var_params['kwargs']
-        # verify the required parameter 'org_id' is set
-        if ('org_id' not in local_var_params or
-                local_var_params['org_id'] is None):
-            raise ValueError("Missing the required parameter `org_id` when calling `post_orgs_id_labels`")  # noqa: E501
-        # verify the required parameter 'label_mapping' is set
-        if ('label_mapping' not in local_var_params or
-                local_var_params['label_mapping'] is None):
-            raise ValueError("Missing the required parameter `label_mapping` when calling `post_orgs_id_labels`")  # noqa: E501
-
-        collection_formats = {}
-
-        path_params = {}
-        if 'org_id' in local_var_params:
-            path_params['orgID'] = local_var_params['org_id']  # noqa: E501
-
-        query_params = []
-
-        header_params = {}
-        if 'zap_trace_span' in local_var_params:
-            header_params['Zap-Trace-Span'] = local_var_params['zap_trace_span']  # noqa: E501
-
-        form_params = []
-        local_var_files = {}
-
-        body_params = None
-        if 'label_mapping' in local_var_params:
-            body_params = local_var_params['label_mapping']
-        # HTTP header `Accept`
-        header_params['Accept'] = self.api_client.select_header_accept(
-            ['application/json'])  # noqa: E501
-
-        # HTTP header `Content-Type`
-        header_params['Content-Type'] = self.api_client.select_header_content_type(  # noqa: E501
-            ['application/json'])  # noqa: E501
-
-        # Authentication setting
-        auth_settings = []  # noqa: E501
-
-        # urlopen optional setting
-        urlopen_kw = None
-        if 'urlopen_kw' in kwargs:
-            urlopen_kw = kwargs['urlopen_kw']
-
-        return self.api_client.call_api(
-            '/api/v2/orgs/{orgID}/labels', 'POST',
-            path_params,
-            query_params,
-            header_params,
-            body=body_params,
-            post_params=form_params,
-            files=local_var_files,
-            response_type='LabelResponse',  # noqa: E501
             auth_settings=auth_settings,
             async_req=local_var_params.get('async_req'),
             _return_http_data_only=local_var_params.get('_return_http_data_only'),  # noqa: E501